Crate mannequin

About

Warning: still under heavy development

This crate contains methods for kinematics and rigging, and dynamics in biomechanics applications and animation. It contains data structures and visitor algorithms for handling tree structures, and will support multiple algebra packages (planned: ndarray, ngalgebra, faer, glam)

See the Mannequin struct to get started.

Reading list

Lecture on Inverse Kinematics

Dynamics in Cumputer Graphics

Naming conventions

Traits – adjectives that indicate capability and behavior (TODO this is currently not always true)

Structs – substantives that indicate entities implementing a behavior

Methods – imperative forms with the exception of getters and factories, which are uses substantives (i.e., omit a get_ prefix) much like the standard library. Callback methods have a on_ prefix

Re-exports

pub use arena::ArenaNode;

pub use arena::ArenaTree;

pub use iterable_tree::Nodelike;

pub use iterable_tree::Order;

pub use iterable_tree::Order::BreadthFirst;

pub use iterable_tree::Order::DepthFirst;

pub use iterable_tree::Order::Unordered;

pub use iterable_tree::TreeIterable;

pub use mannequin::Forward;

pub use mannequin::Inverse;

pub use mannequin::Mannequin;

pub use rigid::accumulate;

pub use rigid::Accumulator;

pub use rigid::Rigid;

Modules

arena

Implementation of a tree iteration with arena allocation

dummy

Dummy implementation of a rigid body used for testing

iterable_tree

mannequin

This module defines the flat, composable architecture for a mannequin (an animated character). Different algorithms for forward and inverse kinematics can be freely combined (as long they are using the same backend) such as rigid bodies, deformable skinning, realistic joints, muscle simulation, or classical inverse kinematics and obstacle avoidance.

rigid

Defines the payload carried by Nodelike in the context of kinematics/character animation