Trait mannequin::rigid::Rigid

pub trait Rigid {
    type Transformation: Clone;
    type Point;
    type Parameter;

    // Required methods
    fn transformation(&self, param: &Self::Parameter) -> Self::Transformation;
    fn globalize(&self, other: &Self::Point) -> Self::Point;
    fn localize(&self, other: &Self::Point) -> Self::Point;
    fn derivative(&self, param: Self::Parameter) -> Self::Transformation;
    fn n_dof(&self) -> i32;
    fn neutral_element() -> Self::Transformation;
    fn concat(
        first: &Self::Transformation,
        second: &Self::Transformation,
    ) -> Self::Transformation;
}

A Rigid Body represents a single, rigid link connected to other links via a joint.

Required Associated Types

type Transformation: Clone

E.g., 4x4 matrix, (3x1, 3x3), quaternions …

type Point

Vec, [f64;4], …

type Parameter

typically joint positions (angles/extension), f64, [f64,3]

Required Methods

fn transformation(&self, param: &Self::Parameter) -> Self::Transformation

Get the Transformation from the parent taking the connecting joint into account

fn globalize(&self, other: &Self::Point) -> Self::Point

Transform a point into the world coordinate system

fn localize(&self, other: &Self::Point) -> Self::Point

Transform a point into the local coordinate system

fn derivative(&self, param: Self::Parameter) -> Self::Transformation

…

fn n_dof(&self) -> i32

…

fn neutral_element() -> Self::Transformation

Returns the eutral element wrt. the transoformation convention used

fn concat( first: &Self::Transformation, second: &Self::Transformation, ) -> Self::Transformation

Concat two transformations

Object Safety

This trait is not object safe.